refactored note templates code
