+++ packages/foam-vscode/src/services/errors.ts
@@ -1,0 +1,8 @@
+export class UserCancelledOperation extends Error {
+  constructor(message?: string) {
+    super('UserCancelledOperation');
+    if (message) {
+      this.message = message;
+    }
+  }
+}
